% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_Rd.R
\name{check_Rd}
\alias{check_Rd}
\title{Checke \code{.Rd} File}
\usage{
check_Rd(path)
}
\arguments{
\item{path}{\link[base]{character} scalar, directory of one \code{.Rd} file}
}
\description{
Checking compliance to some conventions of CRAN.
}
