% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_parent_dir.R
\name{clean_parent_dir}
\alias{clean_parent_dir}
\title{Remove Augment Files in Parent Directory of a Package in Development}
\usage{
clean_parent_dir(pkg = ".")
}
\arguments{
\item{pkg}{..}
}
\description{
Remove Augment Files in Parent Directory of a Package in Development
}
\details{
To remove the augment \verb{*.pdf} and \code{.Rcheck} file/directories created
in the parent directory of a package in development.
}
\examples{
# clean_parent_dir(pkg = '../groupedHyperframe')
}
\keyword{internal}
