% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/document_.R
\name{document_}
\alias{document_}
\title{Auxiliary Tasks of \link[devtools]{document}}
\usage{
document_(pkg = ".")
}
\arguments{
\item{pkg}{\link[base]{character} scalar, directory of the package}
}
\description{
Auxiliary tasks of \link[devtools]{document}, including
\itemize{
\item {checking compliance to some conventions of CRAN}
\item {\link[devtools]{spell_check}}
\item {\link[devtools]{build_manual}}
}
}
