% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/openManual.R
\name{openManual}
\alias{openManual}
\title{Open pdf Manual}
\usage{
openManual(x)
}
\arguments{
\item{x}{returned object from function \link[devtools]{build_manual}}
}
\description{
Open pdf Manual
}
\keyword{internal}
