% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/release_.R
\name{release_}
\alias{release_}
\title{Auxiliary Tasks before \link[devtools]{release}}
\usage{
release_(pkg = ".")
}
\arguments{
\item{pkg}{\link[base]{character} scalar, directory of the package}
}
\description{
Auxiliary tasks before \link[devtools]{release}, including
\itemize{
\item {\link[devtools]{check}}
\item {\code{rhub::rhub_check}; still working on this}
\item {\link[devtools]{check_win_devel}}
}
}
