% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/removeLocalPackage.R
\name{removeLocalPackage}
\alias{removeLocalPackage}
\title{Remove Old Version of a Package}
\usage{
removeLocalPackage(name = ".")
}
\arguments{
\item{name}{\link[base]{character} scalar, name of the package}
}
\description{
Remove old version of a package from \link[base]{search} path,
as well as from \link[utils]{installed.packages}.
}
