% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/updateDESCRIPTION.R
\name{updateDESCRIPTION}
\alias{updateDESCRIPTION}
\title{Update DESCRIPTION File}
\usage{
updateDESCRIPTION(pkg = ".", which = "dev", ...)
}
\arguments{
\item{pkg}{\link[base]{character} scalar, directory of the package}

\item{which}{see function \link[usethis]{use_version}}

\item{...}{additional parameters, currently not in use}
}
\description{
Update DESCRIPTION file
}
\details{
Update the DESCRIPTION file, so that

\describe{

\item{\code{Version}}{using \link[usethis]{use_version} with \code{which = 'dev'}}

\item{\code{Date}}{is today (via \link[base]{Sys.Date})}

}
}
